NEW/simulator/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
openai>=1.0.0
pydantic>=2.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
